Foundation: add some type conversion extensions for Windows
The file attribute flags are used extensively in the file manipulation
code.  Explicitly converting to the value to `DWORD` adds no value and
clutters the code.  Provide an internal overload to translate the type
appropriately.

+++ Sources/Foundation/FileHandle.swift
@@ -880,7 +880,7 @@ extension FileHandle {
                 if error == ERROR_ACCESS_DENIED {
                     var fileInfo = BY_HANDLE_FILE_INFORMATION()
                     GetFileInformationByHandle(self._handle, &fileInfo)
-                    if fileInfo.dwFileAttributes & DWORD(FILE_ATTRIBUTE_DIRECTORY) == DWORD(FILE_ATTRIBUTE_DIRECTORY) {
+                    if fileInfo.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY == FILE_ATTRIBUTE_DIRECTORY {
                         translatedError = Int32(ERROR_DIRECTORY_NOT_SUPPORTED)
                     }
                 }
